refactor(co-parent): tile-driven settlement dialog, drop full-width card
Overview's "Evelina shares" tile becomes the primary surface — shows the
real signed balance with direction (you owe / they owe / settled) and
opens a per-co-parent dialog with the breakdown + Mark-as-settled action.
Standalone yellow card removed. Shared fetch logic extracted into a
useCoParentSettlements hook so the tile and dialog read the same balance.
Bundled in: section header spacing on Overview (leading-none on the h2 +
unified mb-1.5 wrapper) so Recurring and Recent activity sit the same
distance above their cards.
